+++ fr.centralesupelec.edf.riseclipse.iec61850.scl.validator/src/fr/centralesupelec/edf/riseclipse/iec61850/scl/validator/nsd/LNClassValidator.java
@@ -178,29 +178,26 @@ public class LNClassValidator {
             // TODO: which one is the right test?
 //            if( ! nsIdentification.dependsOn( NsIdentification.of( doNamespaces.get( do_.getName() )))) {
             if( ! NsIdentification.of( doNamespaces.get( do_.getName() )).dependsOn( nsIdentification )) {
-                RiseClipseMessage warning = RiseClipseMessage.warning( LNCLASS_VALIDATION_NSD_CATEGORY, do_.getFilename(), do_.getLineNumber(), 
-                        "DO \"", do_.getName(), "\" cannot be validated because its DOI namespace \"", doNamespaces.get( do_.getName() ), "\"is not compatible with the LNClass namespace \"", nsIdentification, "\"" );
+                RiseClipseMessage notice = RiseClipseMessage.notice( LNCLASS_VALIDATION_NSD_CATEGORY, do_.getFilename(), do_.getLineNumber(), 
+                        "DO \"", do_.getName(), "\" namespace \"", doNamespaces.get( do_.getName() ), "\" is different from the LNClass namespace \"", nsIdentification, "\"" );
                 diagnostics.add( new BasicDiagnostic(
-                        Diagnostic.WARNING,
+                        Diagnostic.INFO,
                         RiseClipseValidatorSCL.DIAGNOSTIC_SOURCE,
                         0,
-                        warning.getMessage(),
-                        new Object[] { do_, warning } ));
+                        notice.getMessage(),
+                        new Object[] { do_, notice } ));
                 return res;
             }
-            String[] names;
-            if( do_.getName().matches( "[a-zA-Z]+\\d+" )) {
-                names = do_.getName().split( "(?=\\d)", 2 );
+            
+            // Look for first with the full name, keeping potential ending digits
+            CDCValidator cdcValidator = dataObjectValidatorMap.get( do_.getName() );
+            if( cdcValidator == null ) {
+                if( do_.getName().matches( "[a-zA-Z]+\\d+" )) {
+                    // TODO: a suffix number may be added only when presence condition is Mmulti or Omulti
+                    cdcValidator = dataObjectValidatorMap.get( do_.getName().split( "(?=\\d)", 2 )[0] );
+                }
             }
-            else {
-                names = new String[] { do_.getName() };
-            }
-            if( names.length == 0 ) {
-                // error should have been already displayed
-                //AbstractRiseClipseConsole.getConsole().error( "[NSD validation] Unexpected DO name " + do_.getName() + " in LNodeType (line " + do_.getParentLNodeType().getLineNumber() );
-                continue;
-            }
-            CDCValidator cdcValidator = dataObjectValidatorMap.get( names[0] );
+
             if( cdcValidator != null ) {
                 if(( do_.getRefersToDOType() != null ) && ! cdcValidator.getName().equals( do_.getRefersToDOType().getCdc() )) {
                     RiseClipseMessage warning = RiseClipseMessage.warning( LNCLASS_VALIDATION_NSD_CATEGORY, do_.getFilename(), do_.getLineNumber(),